Refactor to clean arch
