+++ internal/domain/model/anime.go
